pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference."
authors = [
	{ name = "Захарчук Диана", email = "heilig1996@mail.ru" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.9.10",
    "pyyaml==6.0.1",
    "pytest-cov >= 5.0.0",
]
